Introduce derived classes CudaKernel and RocmKernel
This change makes `GpuKernel` an abstract base class
and moves its implementation into the derived classes
`CudaKernel` and `RocmKernel`.

This avoids having two implementations for the same functions
and also reduces the exposure of gpu_types.h which we want to
get rid of.

I'm also adding some basic tests for the new classes.

PiperOrigin-RevId: 678175887